broke transceiver out of the hierarchy to be instantiated at the top
This is more flexible in different kinds of environments. In particular,
it makes using GTP transceivers easier. These transceivers are not
100% independent channels. All transceivers in a quad share 2 PLLs
and the user needs more control over how and where these are
instantiated and connected.

Introduced 'transceiver_pkg.vhd' defining record types to bundle
the incoming and outgoing signals so that they can be routed
around more easily.