Break the transceiver instantiation out of the EVR module
This makes it easier to support multiple/different MGTs.
I.e., GTP devices are closer coupled with the PLLs and
require routing signals between the channels. All of this
can best be handled outside of the EVR proper.

Bundled the relevant signals out of and into the MGT
into EvrTransceiverObType / EvrTransceiverIbType,
respectively.