delay_measure_wnd.vhd: transformed counting process (timing!)
The process which did the actual measurement contained a very long
combinatorial chain (18 stages or so) which was challenging for
timing closure.

I transformed the algorithm so that it does an intermediate computation
into a register (diff_cnt). The goal was not a redesign of the algorithm
or the most elegant solution but the one that was easiest to see that
it produces identical results to the original version.

- sequence of IF statements was transformed into a combinatorially
  equal IF with multiple branches. Individual branches were then
  simplified.
- it could be shown that 'diff_cnt' can essentially be computed
  one cycle ahead (it is updated whenever 'cnt' or 'prev_cnt' are
  updated and)
- since the original logic evaluated
    cnt := cnt + 1
    diff_cnt := prev_cnt - cnt
    if ( diff_cnt ...) then
   we could pre-subtract '1' from the registered 'diff_cnt', i.e.,
   a diff_cnt register is initialized to prev_cnt - 1 and decremented
   every time 'cnt' is incremented. Thus, the above 'if' can then
   operate on a register...